+++ src/init/index.ts
@@ -3,8 +3,8 @@ import { welcome } from "./welcome";
 import { devInit } from "./devInit";
 
 export async function init(context: vscode.ExtensionContext) {
+  welcome(context);
+
   const settings = vscode.workspace.getConfiguration('as2.clients');
-  // TODO: download clientDefaults.json from remote repo
-  await welcome(context);
-  await devInit(settings);
+  devInit(settings);
 }